proxy

+++ automation/updateproxies.sh
@@ -3,7 +3,7 @@
 YELLOW='\033[0;33m'
 RED='\033[0;31m'
 GREEN='\033[0;32m'
-GRAY='\033[0;37m'
+GRAY='\033[0;30m'
 NC='\033[0m' # No Color
 
 repo=https://raw.githubusercontent.com/GiDRaPiTaRdA/DDoSRipper/master
@@ -22,6 +22,8 @@ else
     echo -e "${tmp} ${RED}does not exist or empty${NC}"
 fi
 
-echo -e "${GRAY}remove${NC} $tmp"
-sudo rm $tmp
-
+if [ -e $tmp ]
+then
+    echo -e "${RED}remove${NC} $tmp"
+    sudo rm $tmp
+fi
